+++ src/bar-chart/bar-horizontal.component.ts
@@ -76,7 +76,7 @@ export class BarHorizontalComponent extends BaseChartComponent {
   @Input() schemeType: string;
   @Input() xAxisTickFormatting: any;
   @Input() yAxisTickFormatting: any;
-  @Input() innerPadding = 8;
+  @Input() barPadding = 8;
 
   @Output() activate: EventEmitter<any> = new EventEmitter();
   @Output() deactivate: EventEmitter<any> = new EventEmitter();
@@ -131,7 +131,7 @@ export class BarHorizontalComponent extends BaseChartComponent {
 
   getYScale() {
     this.yDomain = this.getYDomain();
-    const spacing = this.yDomain.length / (this.dims.height / this.innerPadding + 1);
+    const spacing = this.yDomain.length / (this.dims.height / this.barPadding + 1);
     
     return d3.scaleBand()
       .rangeRound([this.dims.height, 0])