+++ src/bar-chart/bar-vertical-normalized.component.ts
@@ -98,7 +98,7 @@ export class BarVerticalNormalizedComponent extends BaseChartComponent {
   @Input() schemeType: string;
   @Input() xAxisTickFormatting: any;
   @Input() yAxisTickFormatting: any;
-  @Input() innerPadding = 8;
+  @Input() barPadding = 8;
 
   @Output() activate: EventEmitter<any> = new EventEmitter();
   @Output() deactivate: EventEmitter<any> = new EventEmitter();
@@ -179,7 +179,7 @@ export class BarVerticalNormalizedComponent extends BaseChartComponent {
   }
 
   getXScale() {
-    const spacing = this.groupDomain.length / (this.dims.width / this.innerPadding + 1);
+    const spacing = this.groupDomain.length / (this.dims.width / this.barPadding + 1);
 
     return d3.scaleBand()
       .rangeRound([0, this.dims.width])