+++ src/bar-chart/bar-vertical.component.ts
@@ -76,7 +76,7 @@ export class BarVerticalComponent extends BaseChartComponent {
   @Input() schemeType: string;
   @Input() xAxisTickFormatting: any;
   @Input() yAxisTickFormatting: any;
-  @Input() innerPadding = 8;
+  @Input() barPadding = 8;
 
   @Output() activate: EventEmitter<any> = new EventEmitter();
   @Output() deactivate: EventEmitter<any> = new EventEmitter();
@@ -123,7 +123,7 @@ export class BarVerticalComponent extends BaseChartComponent {
 
   getXScale() {
     this.xDomain = this.getXDomain();
-    const spacing = this.xDomain.length / (this.dims.width / this.innerPadding + 1);
+    const spacing = this.xDomain.length / (this.dims.width / this.barPadding + 1);
     return d3.scaleBand()
       .rangeRound([0, this.dims.width])
       .paddingInner(spacing)